requirements.txt
Django==5.0.1
pandas==2.1.4
matplotlib==3.8.2
pillow==10.2.0
openai
google-generativeai
huggingface_hub
python-dotenv
numpy
requests 